+++ lib/features/library_character_data/data/models/alignment_model.dart
@@ -1,6 +1,6 @@
 import 'package:json_annotation/json_annotation.dart';
 
-import '../../../app/data/models/base/description_base_model.dart';
+import '../../../../app/data/models/base/description_base_model.dart';
 
 part 'alignment_model.g.dart';
