+++ lib/features/library_character_data/data/models/proficiency_model.dart
@@ -1,6 +1,6 @@
 import 'package:json_annotation/json_annotation.dart';
 
-import '../../../app/data/models/base/reference_base_model.dart';
+import '../../../../app/data/models/base/reference_base_model.dart';
 
 part 'proficiency_model.g.dart';
