+++ lib/features/library_character_data/data/models/skill_model.dart
@@ -1,7 +1,7 @@
 import 'package:json_annotation/json_annotation.dart';
 
-import '../../../app/data/models/base/description_base_model.dart';
-import '../../../app/data/models/base/reference_base_model.dart';
+import '../../../../app/data/models/base/description_base_model.dart';
+import '../../../../app/data/models/base/reference_base_model.dart';
 
 part 'skill_model.g.dart';
